add flag to demangle symbols using llvm-cxxfilt
When the --demangle flag is used, this additional logic will take
effect.

For C++ code we'll need to pass in the mangled versions of
function/method names. This reads the ELF symbols from the binary using
the same method as for the -symbols-prefix flag, but then uses
llvm-cxxfilt to map between the two.

The behavior of --symbols, --symbols-file, and --symbols-prefix are
unaffected.